Refactor semihost and add llvm_libc_errno
Uses a single threaded implementation. As the embedding API is
unstable and may grow, I have made it easier to modify and
code review, by grouping it all together.
