+++ arm-software/embedded/llvmlibc-samples/src/baremetal-semihosting-llvmlibc/hello.c
@@ -10,12 +10,6 @@
 #include <string.h>
 #include <math.h>
 
-// Implementation of errno
-int *__llvm_libc_errno() {
-  static int internal_err;
-  return &internal_err;
-}
-
 // Example that uses heap, string and math library.
 
 int main(void) {
