+++ arm-software/embedded/llvmlibc-support/platform.h
@@ -15,11 +15,19 @@
 #ifndef LLVMET_LLVMLIBC_SUPPORT_PLATFORM_H
 #define LLVMET_LLVMLIBC_SUPPORT_PLATFORM_H
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 // libcrt0.a will call this function after the stack pointer is
 // initialized. If any setup specific to the libc porting layer is
 // needed, this is where to do it. For example, in semihosting, the
 // standard I/O handles must be opened via the SYS_OPEN operation, and
 // this function is where libsemihost.a does it.
 void _platform_init(void);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif // LLVMET_LLVMLIBC_SUPPORT_PLATFORM_H
